day 2: keep iterating if we undershoot the bound
If repeating the prefix is *lower* than the lower bound, then we can keep
iterating through it -- but we also need to filter it out of the sum.

This, too, was found by printing the id range and target lengths and
`.inspect()`ing the iterator through the partial prefixes.  I noticed by eye
that a particular range was suspiciously yielding no invalid IDs which felt
weird for the personal puzzle input...