day 9: start figuring out if it's in the loop
This foray through handedness of a corner is demonstrably wrong:

    #-------A
    |       |
    |       L----C
    |         @  |
    |            |
    #------------#

The point marked @ is on the "left" side of the A->L line, but the right side
of the L->C line, even though it is actually inside the polygon.

I'm going to have to actually look up an algorithm for this.