day 10: model buttons as constraints in microlp
Only minor panic!() as I forgot that the buttons were stored as lists of
indexes and not already expanded into matrices of ones and zeroes.
`.contains()` is linear (so the whole conversion process is quadratic), but I'm
hoping that n is small enough and the constant factor is small enough that I
don't actually have to care.